+++ Changelog.md
@@ -9,6 +9,9 @@
 - Docker
   - add [Supervisor](http://supervisord.org/index.html)
   - add [Composerize](https://github.com/magicmark/composerize)
+  - add [CTop](https://github.com/bcicen/ctop) v0.7.2
+  - add [docker-gc](https://github.com/spotify/docker-gc)
+  - add [containrrr/watchtower](https://github.com/containrrr/watchtower)
   - rTorrent image
 
 ## v7.1 - 2019/11/01 _(started at: 2019/10/02)_